Rip and replace the pattern matching engine
The previous design, while beautiful, was inefficient and not
capable of correctly handling non-greedy matching inside a tree
node. An example of this is with the common pattern for trimming
strings, `^%s*(.-)%s*$`.

A non-greedy match needs to take one token per iteration until
the *entire pattern* fails to parse, but this is not possible when
only a subtree is visible to the repetition function. Trying to
deal with this by naïve linearisation means that the edge of the
capture is lost.

I imagine that there is some technique that would allow this to
work (I am no parser expert) but the thing about Lua patterns is
that they are way too simple for this level of complexity in the
first place. So, this commit deletes the old engine and replaces it
with a hand-translation of the PUC-Lua pattern matching code which
works and passes the Lua test suite.